+++ src/interfaces/UsdnProtocol/IUsdnProtocolTypes.sol
@@ -57,7 +57,7 @@ struct PendingActionCommonData {
     uint40 timestamp; // 5 bytes
     address user; // 20 bytes
     address to; // 20 bytes
-    uint24 securityDepositValue; // 3 bytes
+    uint64 securityDepositValue; // 8 bytes
 }
 
 /**
